services/stream-processor/requirements.txt
kafka-python-ng==2.2.3
redis==5.0.1
psycopg2-binary>=2.9.10